requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
pymongo==4.12.0
python-dotenv==1.1.0
werkzeug==3.1.3
gunicorn==23.0.0